src/requirements.txt
configparser
numpy
pandas
geopandas
matplotlib.pyplot